third_party/blink/web_tests/wpt_internal/app-history/navigate-event/navigate-anchor-with-target.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="iframe" name="i" src="/common/blank.html"></iframe>
<a id="a" href="foo.html" target="i"></a>
<script>
async_test(t => {
  window.onload = t.step_func(() => {
    appHistory.onnavigate = t.step_func_done(() => {
      assert_unreached("onnavigate should not have fired in source window");
    });
    iframe.contentWindow.appHistory.onnavigate = t.step_func_done(e => {
      assert_equals(e.canRespond, true);
      assert_equals(e.hashChange, false);
      assert_equals(e.formData, null);
      e.preventDefault();
    });
    a.click();
  });
}, "<a> with a target fires navigate event in target window but not source");
</script>
